History values fit within `int16_t` (#449)
Bench: 5382108

STC

ELO   | 2.58 +- 4.16 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=8MB
LLR   | 2.96 (-2.94, 2.94) [-4.00, 1.00]
GAMES | N: 12776 W: 3107 L: 3012 D: 6657

+++ src/types.h
@@ -105,7 +105,7 @@ typedef struct {
   Move moves[MAX_SEARCH_PLY];
 } PV;
 
-typedef int PieceTo[12][64];
+typedef int16_t PieceTo[12][64];
 
 typedef struct {
   int ply, staticEval, de;
@@ -165,10 +165,10 @@ struct ThreadData {
   RootMove rootMoves[MAX_MOVES];
   uint64_t nodeCounts[64 * 64];
 
-  Move counters[12][64];     // counter move butterfly table
-  int hh[2][2][2][64 * 64];  // history heuristic butterfly table (stm / threatened)
-  int ch[2][12][64][12][64]; // continuation move history table
-  int caph[6][64][7];        // capture history
+  Move counters[12][64];         // counter move butterfly table
+  int16_t hh[2][2][2][64 * 64];  // history heuristic butterfly table (stm / threatened)
+  int16_t ch[2][12][64][12][64]; // continuation move history table
+  int16_t caph[6][64][7];        // capture history
 
   int action, calls;
   pthread_t nativeThread;
